feat(authentication): add JWT and Identity
This commit introduces JWT (JSON Web Token) and Identity to the
project.The JWT is used for authentication and the Identity is used for
usermanagement. The JWT settings are added to the appsettings.json
file.The Identity is integrated into the ApplicationDBContext and the
necessaryIdentity tables are added to the database schema.

BREAKING CHANGE: This changes the database schema and adds new
dependencies.